+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - SwitchWithLabel
 - Rating
 - Autocomplete
+- Floating Action Button
+- Slider
+
 
 ---