+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 21.11.2022
+
+- Box
+
+---
+
 ## 20.11.2022
 
 - Checkbox
@@ -8,7 +14,6 @@
 - Autocomplete
 - Floating Action Button
 - Slider
-
 
 ---